+++ backend/app/src/main/java/com/yagubogu/leaderboard/service/LeaderboardService.java
@@ -6,10 +6,7 @@ import com.yagubogu.leaderboard.domain.LeaderboardType;
 import com.yagubogu.leaderboard.dto.LeaderboardItemResponse;
 import com.yagubogu.leaderboard.dto.LeaderboardResponse;
 import com.yagubogu.leaderboard.dto.LeaderboardRow;
-import jakarta.annotation.PostConstruct;
-import java.util.EnumMap;
 import java.util.List;
-import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,24 +14,13 @@ import org.springframework.stereotype.Service;
 @Service
 public class LeaderboardService {
 
-    private final Map<LeaderboardType, LeaderboardQuery> map = new EnumMap<>(LeaderboardType.class);
-    private final List<LeaderboardQuery> queries;
-
-    @PostConstruct
-    void init() {
-        for (LeaderboardQuery query : queries) {
-            map.put(query.supports(), query);
-        }
-    }
+    private final LeaderboardQueryRegistry registry;
 
     public List<LeaderboardResponse> findAllTop(int limit) {
         return stream(LeaderboardType.values())
                 .map(type -> {
-                    LeaderboardQuery query = map.get(type);
-                    if (query == null) {
-                        throw new IllegalStateException("No LeaderboardQuery registered for type:" + type);
-                    }
-                    List<LeaderboardItemResponse> items = query.findTop(limit).stream()
+                    List<LeaderboardItemResponse> items = registry.getQuery(type)
+                            .findTop(limit).stream()
                             .map(this::toItem)
                             .toList();
                     return LeaderboardResponse.of(type, items);
